seperating tier from general stuff
